requirements.txt
pyzmq
torch
transformers
pyinstaller